refactor determineRoute to be its own module; move StatusRecord to be its own interface
